+++ module/data/active-effect.js
@@ -61,6 +61,13 @@ export class TorgActiveEffectData extends (foundry.data.ActiveEffectTypeDataMode
             initial: CONST.TOKEN_DISPOSITIONS.FRIENDLY,
             validationError: "must be a value in CONST.TOKEN_DISPOSITIONS"
           }),
+          visibility: new fields.NumberField({
+            required: true,
+            initial: CONST.REGION_VISIBILITY.LAYER_UNLOCKED,
+            choices: Object.values(CONST.REGION_VISIBILITY),
+            label: 'REGION.FIELDS.visibility.label',
+            hint: 'REGION.FIELDS.visibility.hint',
+          }),
         })
       })
     return schema;